+++ src/app/App.css
@@ -8,7 +8,7 @@
 
 .content {
   flex: 1;
-  background-image: url(../../assets/mainBack2.jpg);
+  background-image: url(../assets/mainBack2.jpg);
   background-size: cover;
   background-repeat: no-repeat;
   background-position: right center;
